refactor: Footer 관련 컴포넌트 폴더 구조 변경
# fsd 아키텍처를 참조한 폴더 구조 리팩토링을 진행합니다.
해당 커밋에서 생성된 폴더는 app, widgets, features, entities, shared 폴더 입니다.

- 위젯 단위 (widgets)
  1. Footer: 푸터는 버거풋 전반적으로 사용되는 UI 요소로 widgets 폴더로 이동합니다.

- 기능 단위 (features)
  1. DdayCounter: 특정 날짜까지의 카운트다운을 보여주는 기능이므로, features 폴더로 이동합니다.
  2. UpdatedValueChecker: 데이터를 가져와 업데이트 사항을 보여주는 기능이므로, features 폴더로 이동합니다. 해당 컴포넌트 내에서 함께 사용되는
  3. HowToUse: 사용자에게 간단한 툴팁을 제공하는 기능이므로, features 폴더로 이동합니다.

- 엔티티 단위 (entities)
  1. user와 관련된 zustand 스토어: 사용자 정보를 관리하는 부분으로 entities/user 폴더로 이동합니다.

- 공통 단위 (shared)
  1. Modal: 모달은 Footer를 제외한 다른 컴포넌트 내에서도 재사용이 가능해 shared 폴더로 이동합니다.
  2. BackDrop: 모달과 함께 사용되는 컴포넌트로 Modal 내의 ui 폴더로 이동합니다.
