+++ src/api/Products.js
@@ -1,13 +1,5 @@
 import client from "./index";
 
-// 달라진 목록을 받아오는 api
-export async function getDiffValue() {
-  return await client
-    .get("/loading/result")
-    .then((res) => res.data)
-    .catch((error) => console.log(error));
-}
-
 // 기기 목록을 받아오는 api
 export async function getAllMachines() {
   return await client
