+++ src/entities/user/index.js
@@ -1,0 +1,4 @@
+export { useUserName } from "./model/store";
+export { useSetUserName } from "./model/store";
+export { useUserId } from "./model/store";
+export { useSetUserId } from "./model/store";
