+++ src/features/dday-counter/index.js
@@ -1,0 +1,1 @@
+export { default as DdayCounter } from "./ui/DdayCounter/DdayCounter";
