+++ src/features/how-to-use/index.js
@@ -1,0 +1,1 @@
+export { default as HowToUse } from "./ui/HowToUse/HowToUse";
