+++ src/features/how-to-use/ui/HowToUse/HowToUse.jsx
