+++ src/features/updated-value-check/index.js
@@ -1,0 +1,1 @@
+export { default as UpdatedValueChecker } from "./ui/UpdatedValueChecker/UpdatedValueChecker";
