+++ src/shared/ui/Modal/index.js
@@ -1,0 +1,2 @@
+export { default as Modal } from "./ui/Modal/Modal";
+export { useModal } from "./lib/hooks/useModal";
