+++ src/shared/ui/Modal/lib/hooks/useModal.js
@@ -1,0 +1,17 @@
+import { useState } from "react";
+
+export function useModal(onCloseCallback) {
+  const [openModal, setOpenModal] = useState(false);
+
+  const handleModalOpen = () => {
+    setOpenModal(true);
+  };
+
+  const handleModalClose = () => {
+    setOpenModal(false);
+
+    if (typeof onCloseCallback === "function") onCloseCallback();
+  };
+
+  return { openModal, handleModalOpen, handleModalClose };
+}
