+++ src/widgets/Footer/ui/Footer.module.css
